backend/requirements.txt
BeautifulSoup4
requests
google-generativeai
flask